requirements.txt
# DotNote - Python Dependencies
# Python 3.12 compatible
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.1.8
blinker==1.9.0
itsdangerous==2.2.0
MarkupSafe==3.0.2

luma.led-matrix==1.7.1
luma.core==2.4.2

pillow==11.1.0

gpiozero==2.0.1
RPi.GPIO==0.7.1